+++ R/codecov.R
@@ -130,6 +130,18 @@ codecov <- function(...,
                           repo = Sys.getenv("WERCKER_GIT_REPOSITORY"),
                           commit = commit %||% Sys.getenv("WERCKER_GIT_COMMIT"))
   # ---------
+  # GitLab-CI
+  # ---------
+  } else if (Sys.getenv("CI") == "true" && Sys.getenv("CI_SERVER_NAME") == "GitLab CI") {
+    # http://docs.gitlab.com/ce/ci/variables/README.html
+    slug <- sub(".*/([^/]+/[^/]+)[.]git", "\\1", Sys.getenv("CI_BUILD_REPO"))
+    codecov_url <- paste0(base_url, "/upload/v2") # nolint
+    codecov_query <- list(service = "gitlab",
+                          branch = branch %||% Sys.getenv("CI_BUILD_REF_NAME"),
+                          build = Sys.getenv("CI_BUILD_ID"),
+                          slug = slug,
+                          commit = commit %||% Sys.getenv("CI_BUILD_REF"))
+  # ---------
   # Local GIT
   # ---------
   } else {