+++ R/replace.R
@@ -1,11 +1,10 @@
 #' @useDynLib covr covr_duplicate_
-replacement <- function(name, env = as.environment(-1)) {
-  target_value <- get(name, envir = env)
+replacement <- function(name, env = as.environment(-1), target_value = get(name, envir = env)) {
   if (is.function(target_value) && !is.primitive(target_value)) {
     if (is_vectorized(target_value)) {
       new_value <- target_value
       environment(new_value)$FUN <- trace_calls(environment(new_value)$FUN, name)
-    } else if (memoise::is.memoised(target_value)) {
+    } else if (is.function(target_value) && inherits(target_value, "memoised")) {
       new_value <- target_value
       environment(new_value)$`_f` <- trace_calls(environment(new_value)$`_f`, name)
     } else {